requirements.txt
aiohttp==3.12.15
sqlalchemy==2.0.43
pydantic==2.11.7
environs==14.3.0
asyncpg==0.30.0
pyjwt==2.10.1
werkzeug==3.1.3
